requirements.txt
PyGithub>=2.0.0
pyyaml
pytest